+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 
-- No changes yet.
+- Add LSP completion for fully-qualified type references.
 
 ## [v1.63.0] - 2026-01-06